refactor: RequestMapping의 URI 변경이 apache 패키지에 영향을 주지 않도록 수정
