+++ tomcat/src/main/java/org/apache/catalina/ControllerContainer.java
@@ -2,8 +2,8 @@ package org.apache.catalina;
 
 import java.util.List;
 import org.apache.catalina.exception.ControllerNotFoundException;
-import org.apache.coyote.Controller;
 import org.apache.coyote.Container;
+import org.apache.coyote.Controller;
 import org.apache.coyote.ExceptionController;
 import org.apache.coyote.http11.request.Request;
